requirements.txt
streamlit==1.45.1
tensorflow==2.19.0
pillow==11.2.1
